+++ tsconfig.json
@@ -1,0 +1,9 @@
+{
+  "compilerOptions": {
+    "baseUrl": "./",
+    "paths": {
+      "@pages/*": ["pages/*"],
+      "@type/*": ["types/*"]
+    }
+  }
+}
