+++ types/report-type.ts
@@ -1,0 +1,6 @@
+export enum ReportType {
+    P2P = '#lblP2P',
+    Rolling = "#lblRolling",
+    FixedPeriodic = "#lblFixed",
+    MultipleDate = "//input[@id='rdoBoth']/..",
+}
